refactor(core): move Triple to separate file
Add `isTriple()` type guard

Issue: BG-38773
