+++ modules/core/src/v2/keychains.ts
@@ -10,8 +10,6 @@ import { validateParams } from '../common';
 
 const co = Bluebird.coroutine;
 
-export type Triple<T> = [T, T, T];
-
 export interface Keychain {
   id: string;
   pub: string;
