+++ modules/core/src/v2/triple.ts
@@ -1,0 +1,6 @@
+export type Triple<T> = [T, T, T];
+
+export function isTriple<T>(arr: T[]): arr is Triple<T> {
+  return arr.length === 3;
+}
+
